add approve feature for companies

+++ src/main/java/ensiastjob/controller/DeleteOfferServlet.java
@@ -2,8 +2,9 @@ package ensiastjob.controller;
 
 import ensiastjob.dao.CandidacyDaoImpl;
 import ensiastjob.dao.OfferDaoImpl;
+import ensiastjob.model.Company;
+import ensiastjob.model.Offer;
 
-import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -14,7 +15,7 @@ import java.io.IOException;
 @WebServlet(name = "DeleteOffer", value = "/delete-offer")
 public class DeleteOfferServlet extends HttpServlet {
     @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
         HttpSession session = request.getSession(false);
 
         if (session.getAttribute("member") == null) {
@@ -23,17 +24,26 @@ public class DeleteOfferServlet extends HttpServlet {
             if (session.getAttribute("role").equals("STUDENT")) {
                 response.sendRedirect("/home-student");
             } else if (session.getAttribute("role").equals("COMPANY")) {
+                Company company = (Company) session.getAttribute("company");
+                if (company.isApproved()) {
+                    OfferDaoImpl offerDao = new OfferDaoImpl();
+                    CandidacyDaoImpl candidacyDao = new CandidacyDaoImpl();
 
-                OfferDaoImpl offerDao = new OfferDaoImpl();
-                CandidacyDaoImpl candidacyDao = new CandidacyDaoImpl();
+                    int offerIdDeleted = Integer.parseInt(request.getParameter("offerIdDeleted"));
 
-                int offerIdDeleted = Integer.parseInt(request.getParameter("offerIdDeleted"));
+                    Offer offer = offerDao.getOfferById(offerIdDeleted);
 
-                candidacyDao.deleteAllCandidaciesByOffer(offerIdDeleted);
+                    if (offer.getCompanyId() == company.getCompanyId()) {
 
-                offerDao.deleteOffer(offerIdDeleted);
+                        candidacyDao.deleteAllCandidaciesByOffer(offerIdDeleted);
 
-                response.sendRedirect("/home-company");
+                        offerDao.deleteOffer(offerIdDeleted);
+
+                        response.sendRedirect("/home-company");
+                    }
+                } else {
+                    response.sendRedirect("/not-approved");
+                }
             }
         }
     }
